requirements.txt
Django
whitenoise
azure.storage.blob